move _default_collate_fn_allow_none to ppsci.data.batch_transform to avoid unfound error
